:recycle: refactor UserDetailsForAdmin to IdentityAccessManagement context
