+++ api/src/identity-access-management/domain/models/UserDetailsForAdmin.js
@@ -1,4 +1,4 @@
-import * as localeService from '../services/locale-service.js';
+import * as localeService from '../../../shared/domain/services/locale-service.js';
 
 class UserDetailsForAdmin {
   constructor(
