+++ api/tests/identity-access-management/unit/domain/models/UserDetailsForAdmin_test.js
@@ -1,5 +1,5 @@
-import { UserDetailsForAdmin } from '../../../../src/shared/domain/models/UserDetailsForAdmin.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { UserDetailsForAdmin } from '../../../../../src/identity-access-management/domain/models/UserDetailsForAdmin.js';
+import { expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Domain | Models | UserDetailsForAdmin', function () {
   let localeService;
